requirements.txt
boto3>=1.1.1
csvkit>=0.9.1
invoke>=0.10.1
pandas>=0.16.2
requests>=2.7.0
gunicorn>=19.3.0
cryptography>=1.0
Flask-Cors>=2.1.0
Flask-BasicAuth>=0.2.0
git+https://github.com/jmcarp/sandman@develop
